requirements.txt
pymodbus
pyserial
Pillow
customtkinter
matplotlib
numpy
pytest>=7.0.0
pytest-mock>=3.6.0
